move app initialization functions to a context for reusability
