+++ src/renderer/contexts/OnlineStatus/types.ts
@@ -2,6 +2,11 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 export interface OnlineStatusInterface {
+  appLoading: boolean;
   online: boolean;
+  setAppLoading: (b: boolean) => void;
   setOnline: (b: boolean) => void;
+  handleInitializeApp: () => Promise<void>;
+  handleInitializeAppOffline: () => Promise<void>;
+  handleInitializeAppOnline: () => Promise<void>;
 }
